LangGraph/requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.0.5
langchain-community>=0.0.20
streamlit>=1.29.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
